+++ Code/InputBindingsManager.cs
@@ -22,7 +22,6 @@ namespace ImageOverlay
 
         private readonly ILog _log;
         private readonly Dictionary<string, InputAction> _actions;
-        private readonly Dictionary<string, ProxyBinding> _bindings;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="InputBindingsManager"/> class.
@@ -31,10 +30,6 @@ namespace ImageOverlay
         {
             _log = Mod.Instance.Log;
             _actions = new ();
-            _bindings = new ();
-
-            // Register custom binding.
-            InputSystem.RegisterBindingComposite<ButtonWithThreeModifiers>();
         }
 
         /// <summary>
@@ -51,50 +46,6 @@ namespace ImageOverlay
         /// Ensures an active instance.
         /// </summary>
         internal static void Ensure() => Instance ??= new ();
-
-        /// <summary>
-        /// Harmony prefix for <c>InputManager.SetBindingImpl</c> to implement custom proxy handling.
-        /// </summary>
-        /// <param name="newBinding">New binding.</param>
-        /// <param name="result">Binding result.</param>
-        /// <returns><c>false</c> if the binding update was intercepted here, <c>true</c> otherwise.</returns>
-        [HarmonyPatch(typeof(InputManager), "SetBindingImpl")]
-        [HarmonyPrefix]
-        internal static bool SetBindingImplPrefix(ProxyBinding newBinding, ref ProxyBinding result)
-        {
-            // Check if this is one of ours.
-            if (newBinding.m_MapName.Equals(MapName) && Instance is not null && Instance._bindings.ContainsKey(newBinding.m_ActionName) && Instance._actions.TryGetValue(newBinding.m_ActionName, out InputAction selectedAction))
-            {
-                Instance._log.Info($"updating binding for action {newBinding.m_ActionName}");
-
-                // Apply new binding.
-                selectedAction.ChangeBinding(0).Erase();
-                switch (newBinding.m_Modifiers.Count)
-                {
-                    default:
-                    case 0:
-                        selectedAction.AddBinding(newBinding.path);
-                        break;
-                    case 1:
-                        selectedAction.AddCompositeBinding("ButtonWithOneModifier").With("modifier", newBinding.m_Modifiers[0].m_Path).With("button", newBinding.path);
-                        break;
-                    case 2:
-                        selectedAction.AddCompositeBinding("ButtonWithTwoModifiers").With("modifier1", newBinding.m_Modifiers[0].m_Path).With("modifier2", newBinding.m_Modifiers[1].m_Path).With("button", newBinding.path);
-                        break;
-                    case 3:
-                        selectedAction.AddCompositeBinding("ButtonWithThreeModifiers").With("modifier1", newBinding.m_Modifiers[0].m_Path).With("modifier2", newBinding.m_Modifiers[1].m_Path).With("modifier3", newBinding.m_Modifiers[2].m_Path).With("button", newBinding.path);
-                        break;
-                }
-
-                // Update record.
-                Instance._bindings[newBinding.m_ActionName] = newBinding;
-                result = newBinding;
-
-                return false;
-            }
-
-            return true;
-        }
 
         /// <summary>
         /// Creates a new input action.
@@ -119,8 +70,7 @@ namespace ImageOverlay
 
             // Create action.
             _log.Info($"adding action key for {actionName}");
-            InputActionMap actionMap = InputManager.instance.inputActions.FindActionMap(MapName);
-            InputAction newAction = actionMap.AddAction(actionName, InputActionType.Button, expectedControlLayout: "Button");
+            InputAction newAction = new (actionName, InputActionType.Button);
 
             // Callback.
             if (callback is not null)
@@ -148,29 +98,6 @@ namespace ImageOverlay
 
             // Add to dictionary.
             _actions.Add(actionName, newAction);
-
-            // Crate proxy mapping.
-            ProxyActionMap proxyMap = InputManager.instance.FindActionMap(MapName);
-            List<ProxyModifier> proxyModifiers = new (modifiers.Count);
-            foreach (string modifier in modifiers)
-            {
-                proxyModifiers.Add(new ProxyModifier() { m_Path = modifier });
-            }
-
-            _bindings.Add(actionName, new ProxyBinding()
-            {
-                m_MapName = MapName,
-                m_ActionName = actionName,
-                m_CompositeName = "Keyboard",
-                m_Name = "binding",
-                m_IsRebindable = true,
-                m_AllowModifiers = true,
-                m_CanBeEmpty = true,
-                m_Usage = BindingUsage.Default | BindingUsage.Overlay | BindingUsage.Tool | BindingUsage.CancelableTool | BindingUsage.Editor,
-                m_Modifiers = proxyModifiers,
-                group = "Keyboard",
-                path = path,
-            });
         }
 
         /// <summary>